Implement linear search algorithm

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,7 +10,7 @@ return (-1);
 
 for (i = 0; i < size; i++)
 {
-printf("Value checked array[%zu] = [%d]\n", i, array[i]);
+printf("Value checked array[%lu] = [%d]\n", (unsigned long)i, array[i]);
 if (array[i] == value)
 return (i);
 }
